% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{draw_agent_outcomes}
\alias{draw_agent_outcomes}
\title{Draw outcome for a given agent}
\usage{
draw_agent_outcomes(agent, period = 1)
}
\arguments{
\item{agent}{agent and their outcome scores}

\item{period}{agent time period to draw from}
}
\value{
vector of 0 or 1 indicating success or failure for each agent outcome
}
\description{
Draw outcome for a given agent
}
\examples{
draw_agent_outcomes(generate_agents(num_agents = 1,
                                    outcomes = c("o1", "o2"))[[1]])
}
