% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{draw_agents}
\alias{draw_agents}
\title{Select an agent for a given history of outcomes}
\usage{
draw_agents(
  case_history,
  method = "thompson",
  agents,
  exclude_agent_ids = NULL,
  outcomes,
  outcome_weights = 1,
  norm_type = "O"
)
}
\arguments{
\item{case_history}{history of cases and outcomes}

\item{method}{method of drawing agent ("random", "oracle", or "thompson")}

\item{agents}{agent matrix}

\item{exclude_agent_ids}{which agent_ids to exclude from selection}

\item{outcomes}{character vector of outcomes to consider}

\item{outcome_weights}{weights to apply to each outcome}

\item{norm_type}{type of norm. Same as \emph{type} in \link[base]{norm}}
}
\value{
name of agent selected
}
\description{
Select an agent for a given history of outcomes
}
