% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_agents}
\alias{generate_agents}
\title{Generate agents with known performance}
\usage{
generate_agents(
  num_agents,
  outcomes,
  num_periods = 1,
  min_val = 0.25,
  max_val = 0.75,
  max_len = 5
)
}
\arguments{
\item{num_agents}{number of agents to generate}

\item{outcomes}{character vector of outcomes this agent produces}

\item{num_periods}{number of different periods of varyiing performance}

\item{min_val}{minimum performance value (expressed as a probability)}

\item{max_val}{maximum performance value (expressed as a probability)}

\item{max_len}{maximum character length of identifiers}
}
\value{
a matrix of performance scores for each agent/time period
}
\description{
Generate agents with known performance
}
\examples{
generate_agents(num_agents = 10, outcomes = c("outcome1", "outcome2"))
}
