% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_cases}
\alias{generate_cases}
\title{Generate a case load}
\usage{
generate_cases(
  num_cases,
  time_period = as.difftime(365, units = "days"),
  start_dt = Sys.time() - time_period,
  end_dt = Sys.time(),
  max_len = 5
)
}
\arguments{
\item{num_cases}{number of cases to generate}

\item{time_period}{period of time that the cases will span}

\item{start_dt}{earliest case date time}

\item{end_dt}{latest case date time}

\item{max_len}{maximum character length of identifiers}
}
\value{
data frame with date_time, case_id, subject_id and characteristic columns
}
\description{
Generate a case load
}
\examples{
generate_cases(n = 100)
}
