% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate random identifiers}
\usage{
generate_ids(n, max_len = 5)
}
\arguments{
\item{n}{number of identifiers to generate}

\item{max_len}{max character length of each identifier}
}
\value{
character vector of identifiers
}
\description{
Generate random identifiers
}
\examples{
generate_ids(n = 10, max_len = 8)
}
