% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{get_max_agent}
\alias{get_max_agent}
\title{Get the agent with the maximum score}
\usage{
get_max_agent(agents, outcome_weights = 1, norm_type = "O")
}
\arguments{
\item{agents}{matrix of agents and outcome scores}

\item{outcome_weights}{weights to apply to each outcome}

\item{norm_type}{type of norm. Same as \emph{type} in \link[base]{norm}}
}
\value{
agent name (character) of agent with max score
}
\description{
Get the agent with the maximum score
}
\examples{
get_max_agent(generate_agents(num_agents = 2, outcomes = c("o1", "o2")))
}
