% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_agent_perf}
\alias{plot_agent_perf}
\title{Plot animated agent performance}
\usage{
plot_agent_perf(
  simulated_cases,
  type = "density",
  outcomes = NULL,
  methods = NULL,
  period = 10,
  x_label = NULL,
  y_label = NULL,
  ...
)
}
\arguments{
\item{simulated_cases}{cases after simulation}

\item{type}{"density" or "scatter" plot}

\item{outcomes}{outcomes to plot}

\item{methods}{character vector of methods of simulation to plot}

\item{period}{animate every "period" cases}

\item{x_label}{label for x-axis}

\item{y_label}{label for y-axes}

\item{...}{options passed to \link[ggplot2]{theme}}
}
\value{
animated plotly plot
}
\description{
Plot animated agent performance
}
