% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot animated agent assignment and outcomes}
\usage{
plot_counts(
  simulated_cases,
  outcomes = NULL,
  methods = NULL,
  agent_scores = NULL,
  period = 10,
  x_label = NULL,
  y_label = NULL,
  color_palette = "Dark2",
  ...
)
}
\arguments{
\item{simulated_cases}{cases after simulation}

\item{outcomes}{outcomes to plot}

\item{methods}{character vector of methods of simulation to plot}

\item{agent_scores}{vector of agents and their scores}

\item{period}{animate every "period" cases}

\item{x_label}{label for x-axis}

\item{y_label}{label for y-axes}

\item{color_palette}{to be used to differentiate colors}

\item{...}{options passed to \link[ggplot2]{theme}}
}
\value{
animated plotly plot
}
\description{
Plot animated agent assignment and outcomes
}
