% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{score_agents}
\alias{score_agents}
\title{Score a set of agents based on their known performance}
\usage{
score_agents(agents, outcome_weights = 1, norm_type = "O")
}
\arguments{
\item{agents}{matrix of agents and outcome scores}

\item{outcome_weights}{weights to apply to each outcome}

\item{norm_type}{type of norm. Same as \emph{type} in \link[base]{norm}}
}
\value{
vector of numeric scores for each agent
}
\description{
Score a set of agents based on their known performance
}
\examples{
score_agents(generate_agents(num_agents = 2, outcomes = c("o1", "o2")))
}
