% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{score_outcomes}
\alias{score_outcomes}
\title{Score a vector of outcomes}
\usage{
score_outcomes(
  outcomes,
  outcome_weights = rep(1, length(outcomes)),
  norm_type = "O"
)
}
\arguments{
\item{outcomes}{numeric vector of outcomes to score}

\item{outcome_weights}{weights to apply to each outcome}

\item{norm_type}{type of norm. Same as \emph{type} in \link[base]{norm}}
}
\value{
a numeric score for the vector of outcomes
}
\description{
Score a vector of outcomes
}
\examples{
score_outcomes(c(3, 4))
}
