% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run a simulation with a set of cases and agents}
\usage{
simulate(
  cases,
  agents,
  outcomes = colnames(agents[[1]]),
  outcome_weights = rep(1, length(outcomes)),
  norm_type = "O",
  methods = c("thompson", "random", "oracle"),
  fixed_agent_list = NULL,
  max_cases = 1,
  max_period = as.difftime(1, units = "days"),
  memory = as.difftime(365, units = "days")
)
}
\arguments{
\item{cases}{case load to simulate}

\item{agents}{matrix of agents and their outcome scores}

\item{outcomes}{character vector of outcomes}

\item{outcome_weights}{weights to apply to each outcome}

\item{norm_type}{type of norm. Same as \emph{type} in \link[base]{norm}}

\item{methods}{methods of drawing agent to simulate ("random", "oracle", "thompson", "fixed")}

\item{fixed_agent_list}{list of fixed agent assignments to simulate}

\item{max_cases}{maximum number of cases an agent can have within max_period}

\item{max_period}{period to consider individual agent case load}

\item{memory}{how many cases to consider when assessing agent performance}
}
\value{
cases annotated with agent selections and outcomes
}
\description{
Run a simulation with a set of cases and agents
}
